pyproject.toml
[tool.poetry]
name = "orkg-nlp-api"
version = "0.3.1"
description = "REST API to ORKG-NLP"
authors = ["Omar Arab Oghli <omar.araboghli@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.10"
fastapi = "~0.76.0"
gunicorn = "~20.1.0"
uvicorn = "~0.17.6"
orkgnlp = "0.9.0"
pytest = "^7.1.2"
python-multipart = "^0.0.5"
tabula-py = "2.4.0"
bert-extractive-summarizer = "0.10.1"
transformers = "4.19.3"
sacremoses = "0.0.53"
SQLAlchemy = "^1.4.39"
psycopg2-binary = "2.9.3"
python-dotenv = "0.20.0"
pre-commit = "^3.1.0"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
